Split Post into PostOverview and PostDetails
There are two views showing posts in the app - a list of all posts and a single post. In the list of all posts we must show less details, and in the view with a single post - the contrary - we must show more details. Before one component was used in the two cases and in order to make the differentiation in the two cases possible - we splitted it into two components.

Also fix name of propType and fix the rendering of post when userData is not available.